Ajustes valor @Root em algumas classes do IBS/CBS
