+++ src/main/java/com/fincatto/documentofiscal/cte400/classes/nota/CTeNotaInfoInformacoesRelativasImpostosIBSCBS.java
@@ -342,4 +342,34 @@ public class CTeNotaInfoInformacoesRelativasImpostosIBSCBS extends DFBase {
       }
     }
   }
+
+  @Root(name = "gEstornoCred")
+  @Namespace(reference = CTeConfig.NAMESPACE)
+  public class CTeTTribEstornoCred extends DFBase {
+
+    private static final long serialVersionUID = -3330020091023450254L;
+
+    @Element(required = true)
+    private String vIBSEstCred;
+
+    @Element(required = true)
+    private String vCBSEstCred;
+
+    public String getVIBSEstCred() {
+      return vIBSEstCred;
+    }
+
+    public void setVIBSEstCred(BigDecimal vIBSEstCred) {
+      this.vIBSEstCred = DFBigDecimalValidador.tamanho13Com2CasasDecimais(vIBSEstCred, "Valor do IBS a ser estornado");
+    }
+
+    public String getvCBSEstCred() {
+      return vCBSEstCred;
+    }
+
+    public void setVCBSEstCred(BigDecimal vCBSEstCred) {
+      this.vCBSEstCred = DFBigDecimalValidador.tamanho13Com2CasasDecimais(vCBSEstCred, "Valor do CBS a ser estornado");
+    }
+
+  }
 }
