+++ src/main/java/com/fincatto/documentofiscal/cte400/classes/nota/CTeTDevTrib.java
@@ -13,7 +13,7 @@ import org.simpleframework.xml.Root;
  *
  * @author Edivaldo Merlo Stens
  */
-@Root(name = "TDevTrib")
+@Root(name = "gDevTrib")
 @Namespace(reference = CTeConfig.NAMESPACE)
 public class CTeTDevTrib extends DFBase {
